Presentation File
This is the PowerPoint presentation file that we presented during the viva examination. We have clearly explained and outlined the functionalities in this file. For those planning to use this project, I recommend exploring this presentation before making any changes to the project source. I would like to credit my team members—Mr. Dilshan, Mr. Sasinda, and Mr. Nawod (Rosa)—for their contributions to the presentation.